backend/data-server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
requests==2.31.0
websockets==12.0
python-dotenv==1.0.0
pydantic==2.5.0
asyncio-mqtt==0.16.1
aiofiles==23.2.1
httpx==0.25.2